feature/ game is fully playable
