+++ index.js
@@ -1,3 +1,8 @@
 // This is where your project starts.
+const {start} = require('./src/game');
+
+
+start()
+
 
 console.log('Your project is running...');
